Split the implementation to several files.
- The `process` object is the only injected object from bin wrapper.
- Extract the `process` from `stdin` module.
